refactor: split tmux activity hysteresis into its own file
app_tmux_activity.go sat at 499/500 lines, one under the hard cap, mixing the activity scan-scheduling lifecycle with the per-session liveness/demotion hysteresis. Move the hysteresis unit (activityDemotionMissThreshold, recordSessionMiss, syncActivitySessionStates, tabSessionInfoByName) into app_tmux_activity_hysteresis.go along the natural A/B seam already recognized by the existing app_tmux_activity_hysteresis_test.go, so the next maintainer touching hysteresis is not forced to split under the cap. No behavior change. Part of the quality stacked diff (stacked on roadmap/07-make-dev-requires-air-with-zero-install-guidan).
